[MAIN GROUP 9] Adiciona Request de produtos + ajustes

+++ back-end/src/database/migrations/20220531221624-create-products.js
@@ -15,7 +15,7 @@ module.exports = {
         unique: true,
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(4,2),
         allowNull: false,
       },
       url_image: {
